+++ pxdlib/errors.py
@@ -1,0 +1,19 @@
+class PixelmatorError(Exception):
+    '''
+    Some change was made which is invalid
+    in a Pixelmator document.
+    '''
+
+
+class ChildError(PixelmatorError):
+    '''
+    Attempted move of a layer to a parent layer
+    for which an invalid error would occur.
+    '''
+
+
+class MaskError(ChildError):
+    '''
+    A layer was moved about in a way that creates
+    a situation where a layer has a mask (but shouldn't)
+    '''