+++ docs/changelog.md
@@ -5,10 +5,10 @@
 ### 0.0.4
 
 - Various bug fixes.
+- Added `PixelmatorError`, and tightened up error raising, in particular with `layer.mask`.
 - Improved `repr(layer)`.
 - Allowed very limited creation of new layers eg `GroupLayer(parent)`
 - Added `layer.delete` and `layer.copyto(parent)`, and thereby added the ability to set `layer.parent` to be in a different `PXDFile`.
-- Tightened up `layer.mask` behaviour.
 
 ### 0.0.3